+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### This file contains changes and modifications during developing.
 
+## Feb 13 2021, 10:18 AM
+- Add 'data_reading' module to read, initially preprocess and decompose documents into paragraphs
+
 ## Feb 3 2021, 11:00 AM
 - Very first commit
 - Add default files: `.gitignore`, `CHANGELOG.md`, `configs.py`, `launcher.sh` and `README.md`